requirements.txt
anyio==3.6.1
bcrypt==4.0.0
black==22.8.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.82.0
fastapi-pagination==0.10.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
protobuf==3.20.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==6.1.2
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.16.1
websockets==10.3
